# Refactor bulk insert for SplitRowStore