This code refactors out multiple calls to the catalog in tight insert loops.

**Correctness**
I did not add any new tests because the tests in SplitRowStore_unittest covered both bulk and bulkWithRemappedAttributes insert methods.

**Benchmarks**
*Process*
Create a table with 6 int attributes and two varchar attributes. Then run a query with 50% selectivity and which selects on the varchar attribute as well as some of the int attributes, and insertion time should be slightly faster. Data in the relation is uniform-random distributed. The machine I tested this on was my laptop which should be a fair canidate because the relation in question is 1GB which easily fits in my memory.
```SQL
-- relation is ~ 1GB
create table splitrow (i0 int,i1 int,i2 int,i3 int,i4 int,s0 varchar(32),s1 varchar(32));
copy splitrow from 'splitrow.qs' WITH (DELIMITER '|');

-- test 1
select * from splitrow where i0 < 5000000; -- half the 1gb relation
select * from splitrow where i0 < 5000000;
select * from splitrow where i0 < 5000000;

-- test 2
select * from splitrow where i0 < 10; -- less than a one thousandth the 1gb relation
select * from splitrow where i0 < 10;
select * from splitrow where i0 < 10;
```

Test 1 results
```
Without change:   16420.618 ms
With change:       8122.209 ms
```

Test 2 results
```
Without change:   603.432 ms
With change:      604.629 ms
```
TLDR

We see a 2x improvement on large inserts.