@navsan 2016-09-20
Please address the comments: mostly to do with performance improvements. Note that the objective here is not just to make SplitRowStore faster than it is today, but to get it to the point where we can drop PackedRowStore. We're not close to that right now, because we're doing far too much work in the inner loop. 

Also, 
- Consider getting rid of occupancy_bitmap_
- Consider moving the null_bitmap into the header as a single bitmap for the entire tuple storage subblock. 
- Consider getting rid of TypedValue