@cramja 2016-09-23
@navsan I made the following changes:
* Make per-column insert info
* Moved some variables outside the loop
* sped up the way that the block estimates the number of tuples it can insert

After these changes, the experiment (mentioned in the header to this PR) speeds up to `7267.457 ms` with the same setup.