@pateljm 2016-10-01
@cramja This is a nice feature. Thanks @navsan for the detailed comments. 

Both, can we close this PR out over the next few days. Gotta keep moving ™ :)