@cramja 2016-10-02
Agreed. I've been working on adding @navsan 's idea of merging contiguous attributes. I have another [branch with those changes](https://github.com/cramja/incubator-quickstep/tree/splitrow_partial_ins). It's almost ready, I'm in the process of benchmarking with TPCH to see if it can replace PackedRow.

Granted this PR speeds up Splitrow, and has tests so can be merged.