@cramja 2016-10-05
I'm closing this PR in favor of another which uses a technique to merge attribute copies when possible.